Refactor dimBungieService into its own module, with split services for UserService and Destiny1Service.
