+++ src/scripts/bungie-api/bungie-api-utils.js
