+++ src/scripts/bungie-api/destiny1-api.service.js
@@ -1,136 +1,31 @@
-import angular from 'angular';
 import _ from 'underscore';
 import { bungieApiQuery, bungieApiUpdate } from './bungie-api-utils';
 
-angular.module('dimApp')
-  .factory('dimBungieService', BungieService);
+/**
+ * APIs for interacting with Destiny 1 game data.
+ */
+export function Destiny1Api(
+  BungieServiceHelper,
+  $q,
+  $http,
+  dimState,
+  $translate) {
+  'ngInject';
+  const { handleErrors, retryOnThrottled } = BungieServiceHelper;
 
-function BungieService($rootScope, $q, $timeout, $http, $state, dimState, $translate) {
-  const service = {
-    getAccounts: getAccounts,
-    getAccountsForCurrentUser: getAccountsForCurrentUser,
-    getCharacters: getCharacters,
-    getStores: getStores,
-    transfer: transfer,
-    equip: equip,
-    equipItems: equipItems,
-    setItemState: setItemState,
-    getManifest: getManifest,
-    getVendorForCharacter: getVendorForCharacter
+  return {
+    getCharacters,
+    getStores,
+    transfer,
+    equip,
+    equipItems,
+    setItemState,
+    getManifest,
+    getVendorForCharacter
   };
-
-  return service;
-  function handleErrors(response) {
-    if (response.status === -1) {
-      return $q.reject(new Error($translate.instant('BungieService.NotConnected')));
-    }
-    // Token expired and other auth maladies
-    if (response.status === 401 || response.status === 403) {
-      $rootScope.$broadcast('dim-no-token-found');
-      return $q.reject(new Error($translate.instant('BungieService.NotLoggedIn')));
-    }
-    if (response.status >= 503 && response.status <= 526 /* cloudflare */) {
-      return $q.reject(new Error($translate.instant('BungieService.Down')));
-    }
-    if (response.status < 200 || response.status >= 400) {
-      return $q.reject(new Error($translate.instant('BungieService.NetworkError', {
-        status: response.status,
-        statusText: response.statusText
-      })));
-    }
-
-    const errorCode = response.data ? response.data.ErrorCode : -1;
-
-    // See https://github.com/DestinyDevs/BungieNetPlatform/wiki/Enums#platformerrorcodes
-    switch (errorCode) {
-    case 1: // Success
-      return response;
-    case 1627: // DestinyVendorNotFound
-      return $q.reject(new Error($translate.instant('BungieService.VendorNotFound')));
-    case 2106: // AuthorizationCodeInvalid
-    case 2108: // AccessNotPermittedByApplicationScope
-      $rootScope.$broadcast('dim-no-token-found');
-      return $q.reject("DIM does not have permission to perform this action.");
-    case 5: // SystemDisabled
-      return $q.reject(new Error($translate.instant('BungieService.Maintenance')));
-    case 35: // ThrottleLimitExceededMinutes
-    case 36: // ThrottleLimitExceededMomentarily
-    case 37: // ThrottleLimitExceededSeconds
-      return $q.reject(new Error($translate.instant('BungieService.Throttled')));
-    case 2111: // token expired
-    case 99: // WebAuthRequired
-      $rootScope.$broadcast('dim-no-token-found');
-      return $q.reject(new Error($translate.instant('BungieService.NotLoggedIn')));
-    case 1601: // DestinyAccountNotFound
-    case 1618: // DestinyUnexpectedError
-      if (response.config.url.indexOf('/Account/') >= 0 &&
-          response.config.url.indexOf('/Character/') < 0) {
-        const error = new Error($translate.instant('BungieService.NoAccount', { platform: dimState.active.label }));
-        error.code = errorCode;
-        return $q.reject(error);
-      }
-    case 2101: // ApiInvalidOrExpiredKey
-    case 2102: // ApiKeyMissingFromRequest
-    case 2107: // OriginHeaderDoesNotMatchKey
-      if ($DIM_FLAVOR === 'dev') {
-        $state.go('developer');
-        return $q.reject(new Error($translate.instant('BungieService.DevVersion')));
-      } else {
-        return $q.reject(new Error($translate.instant('BungieService.Difficulties')));
-      }
-    }
-
-    // Any other error
-    if (errorCode > 1) {
-      if (response.data.Message) {
-        const error = new Error(response.data.Message);
-        error.code = response.data.ErrorCode;
-        error.status = response.data.ErrorStatus;
-        return $q.reject(error);
-      } else {
-        return $q.reject(new Error($translate.instant('BungieService.Difficulties')));
-      }
-    }
-
-    return response;
-  }
-
-  function retryOnThrottled(response, retries = 3) {
-    // TODO: these different statuses suggest different backoffs
-    if (response.data &&
-        // ThrottleLimitExceededMinutes
-        (response.data.ErrorCode === 35 ||
-         // ThrottleLimitExceededMomentarily
-         response.data.ErrorCode === 36 ||
-         // ThrottleLimitExceededSeconds
-         response.data.ErrorCode === 37)) {
-      if (retries <= 0) {
-        return response;
-      } else {
-        return $timeout(Math.pow(2, 4 - retries) * 1000)
-          .then(() => $http(response.config))
-          .then((response) => retryOnThrottled(response, retries - 1));
-      }
-    } else {
-      return response;
-    }
-  }
 
   function getManifest() {
     return $http(bungieApiQuery('/D1/Platform/Destiny/Manifest/'))
-      .then(handleErrors, handleErrors)
-      .then((response) => response.data.Response);
-  }
-
-  function getAccounts(bungieMembershipId) {
-    return $http(bungieApiQuery(`/Platform/User/GetMembershipsById/${bungieMembershipId}/254/`))
-      .then(handleErrors, handleErrors)
-      .then((response) => response.data.Response);
-  }
-
-  // This is here just for migrating folks to GetMembershipsById
-  function getAccountsForCurrentUser() {
-    return $http(bungieApiQuery(`/Platform/User/GetMembershipsForCurrentUser/`))
       .then(handleErrors, handleErrors)
       .then((response) => response.data.Response);
   }
@@ -330,7 +225,7 @@ function BungieService($rootScope, $q, $timeout, $http, $state, dimState, $trans
       .then((response) => {
         const data = response.data.Response;
         store.updateCharacterInfoFromEquip(data.summary);
-        return _.select(items, (i) => {
+        return items.filter((i) => {
           const item = _.find(data.equipResults, {
             itemInstanceId: i.id
           });
