+++ src/scripts/services/dimItemService.factory.js
@@ -5,7 +5,7 @@ angular.module('dimApp')
   .factory('dimItemService', ItemService);
 
 function ItemService(dimStoreService,
-                     dimBungieService,
+                     Destiny1Api,
                      dimCategory,
                      $q,
                      $translate) {
@@ -27,7 +27,7 @@ function ItemService(dimStoreService,
   };
 
   function setItemState(item, store, lockState, type) {
-    return dimBungieService.setItemState(item, store, lockState, type)
+    return Destiny1Api.setItemState(item, store, lockState, type)
       .then(() => lockState);
   }
 
@@ -253,7 +253,7 @@ function ItemService(dimStoreService,
       if (items.length === 1) {
         return equipItem(items[0]);
       }
-      return dimBungieService.equipItems(store, items)
+      return Destiny1Api.equipItems(store, items)
         .then((equippedItems) => {
           return equippedItems.map((i) => {
             return updateItemModel(i, store, store, true);
@@ -266,7 +266,7 @@ function ItemService(dimStoreService,
     if ($featureFlags.debugMoves) {
       console.log('Equip', item.name, item.type, 'to', dimStoreService.getStore(item.owner).name);
     }
-    return dimBungieService.equip(item)
+    return Destiny1Api.equip(item)
       .then(() => {
         const store = dimStoreService.getStore(item.owner);
         return updateItemModel(item, store, store, true);
@@ -299,7 +299,7 @@ function ItemService(dimStoreService,
     if ($featureFlags.debugMoves) {
       console.log('Move', amount, item.name, item.type, 'to', store.name, 'from', dimStoreService.getStore(item.owner).name);
     }
-    return dimBungieService.transfer(item, store, amount)
+    return Destiny1Api.transfer(item, store, amount)
       .then(() => {
         const source = dimStoreService.getStore(item.owner);
         const newItem = updateItemModel(item, source, store, false, amount);
